+++ src/blob/index/bptree/core.rs
@@ -1,3 +1,5 @@
+use crate::error::ValidationErrorKind;
+
 /// structure of b+-tree index file from the beginning:
 /// 1. Header
 /// 2. b+-tree user buffer (now Bloom filter is stored as this buffer)
@@ -122,11 +124,15 @@ impl FileIndexTrait for BPTreeFileIndex {
 
     fn validate(&self) -> Result<()> {
         // FIXME: check hash here?
-        if self.header.is_written() && self.header.version() == HEADER_VERSION {
-            Ok(())
-        } else {
-            Err(Error::validation("Index Header is not valid").into())
+        if !self.header.is_written() {
+            let param = ValidationErrorKind::IndexIsWritten;
+            return Err(Error::validation(param, "Index Header version is not valid").into());
         }
+        if self.header.version() != HEADER_VERSION {
+            let param = ValidationErrorKind::IndexVersion;
+            return Err(Error::validation(param, "Index Header version is not valid").into());
+        }
+        Ok(())
     }
 }
 
@@ -314,11 +320,9 @@ impl BPTreeFileIndex {
 
     async fn validate_header(&self, buf: &mut Vec<u8>) -> Result<()> {
         self.validate()?;
-        if self.header.version() != HEADER_VERSION {
-            return Err(Error::validation("header version mismatch").into());
-        }
         if !Self::hash_valid(&self.header, buf)? {
-            return Err(Error::validation("header hash mismatch").into());
+            let param = ValidationErrorKind::IndexChecksum;
+            return Err(Error::validation(param, "header hash mismatch").into());
         }
         Ok(())
     }